before solving api-tokens

+++ app/Http/Controllers/IncidenceController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Incidencia;
+use App\Models\Zona;
 use Illuminate\Support\Facades\Auth;
 
 class IncidenceController extends Controller
@@ -23,7 +24,8 @@ class IncidenceController extends Controller
      */
     public function create()
     {
-        return view('incidencies.create');
+        $zones = Zona::all();
+        return view('incidencies.create', ['zones' => $zones]);
     }
 
     /**
@@ -36,13 +38,18 @@ class IncidenceController extends Controller
             'Descripcio' => 'required|max:500',
             'Estat' => 'required',
             'Prioritat' => 'required',
-            //'Zona' => 'required',
-            //'imatge' => 'required|image',
+            'Zona' => 'required',
+            'Imatge' => 'required|image',
             'Rol_assignat' => 'required',
         ]);
         $incident = new Incidencia($validatedData);
         //$incident->Usuari_denunciant = Auth::id();
         $incident->Data = now();
+        $photoName = time() . '_' . $request->Imatge->getClientOriginalName();
+        $path = public_path('img/');
+        $request->Imatge->move($path, $photoName);
+        $incident->Ruta_img = $photoName;
+
         // if ($request->hasFile('image')) {
         //     $incident->image = $request->file('image')->store('public/incidences');
         // }
@@ -71,7 +78,11 @@ class IncidenceController extends Controller
 
     public function edit(Incidencia $incidence)
     {
-        return view('incidencies.edit', ['incidencia' => $incidence]);
+        $user = Auth::user();
+        $this->authorize('access', $user);
+        
+        $zones = Zona::all();
+        return view('incidencies.edit', ['incidencia' => $incidence], ['zones' => $zones]);
     }
 
 
@@ -82,18 +93,23 @@ class IncidenceController extends Controller
 
     public function update(Request $request, Incidencia $incidence)
     {
+        $user = Auth::user();
+        $this->authorize('access', $user);
 
         $validatedData = $request->validate([
             'Nom' => 'required|max:150',
             'Descripcio' => 'required|max:500',
             'Estat' => 'required',
             'Prioritat' => 'required',
-            //'Zona' => 'required',
-            //'imatge' => 'required|image',
+            'Zona' => 'required',
             'Rol_assignat' => 'required',
         ]);
-
-
+        if(!is_null($request->Imatge)){
+            $photoName = time() . '_' . $request->Imatge->getClientOriginalName();
+            $path = public_path('img/');
+            $request->Imatge->move($path, $photoName);
+            $incidence->Ruta_img = $photoName;
+        }
         $incidence->update($validatedData);
 
         return redirect()->route('incidences.index')->with('success', 'Incidence updated successfully.');
@@ -107,6 +123,9 @@ class IncidenceController extends Controller
 
     public function destroy(Incidencia $incidence)
     {
+        $user = Auth::user();
+        $this->authorize('access', $user);
+
         $incidence->delete();
 
         return redirect()->route('incidences.index')->with('success', 'Incidence deleted successfully.');
